Add support for non Mercator projection in OL view
give the opportunity to translate the zoom level before giving it to MapLibre. This makes it possible to have OL functioning with a non-Mercator projection system, and still be able to have MapLibre receiving a valid Mercator zoom level to keep the maps in sync.

You can see this code in action here : https://github.com/geoadmin/web-mapviewer/pull/981
where we have our OL instance working in LV95.